fix : change controller logic
- 전체 흐름을 주관하는 책임을 다시 부여
- 이에 따른 테스트 로직 변경
- 테스트에서 mock 제거. 최대한 순수 자바를 사용할 수 있도록 변경
